Business Intelligence Ecommerce sales & Prophet
Analysis of the online sales made by a technology ecommerce of its two categories of phones - PremiumPlus and Affordable. And time series forecast with Prophet.
The data comprises the store's sales from January 2020 to December 2022, including the average price of each category in its respective month. The following analysis will check the number of units sold by each category on a monthly basis, which one makes the most money, its average price and the seasonality of each category. Likewise, a prediction will be made for the first 3 months of the year 2023. 
The data have been slightly modified to maintain confidentiality.